% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add a (drat) repository to the current session}
\usage{
add(name)
}
\arguments{
\item{name}{GitHub account name for which a \sQuote{drat} archive
is to be added.}
}
\value{
The altered set of repositories
}
\description{
R can use multiple archives: CRAN, BioConductor and Omegahat have
been supported for years.  It is equally easy to add local
archives from the same machine, or local network, or university /
company network as well as other publically available
repositories.  This function aids in the process, and defaults to
adding a \sQuote{drat} archive at GitHub under the given name.
}
\details{
This function retrieves the current set of repositories and adds
(or overwrites) the entry for the given \sQuote{name}.
}
\author{
Dirk Eddelbuettel
}

