% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\docType{data}
\name{fips}
\alias{fips}
\title{fips dataset}
\format{
An object of class \code{data.frame} with 3232 rows and 3 columns.
}
\usage{
data(fips)
}
\description{
a fips dataset with 3 columns: FIPS codes, county names, state names
}
\examples{
data(fips)

}
\keyword{datasets}
