% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insertPackage.R
\name{getPathForMac}
\alias{getPathForMac}
\title{Extract R version and built architecture from package DESCRIPTION}
\usage{
getPathForMac(rversion, file)
}
\arguments{
\item{rversion}{String which identifies the major.minor
R version, which was used to build this package. Defaults
to the version of the current interpreter.}

\item{file}{The fully qualified path to the package archive file}
}
\value{
string Relative file path where packages of
this type should be copied to.
}
\description{
This function parses the Built metadata from DESCRIPTION
in a package submission to determine the MacOS path for
binary builds.
}
\author{
Matthew Jones
}

