% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insertPackage.R
\name{getPathForPackage}
\alias{getPathForPackage}
\title{Get relative path for package type}
\usage{
getPathForPackage(file)
}
\arguments{
\item{file}{The fully qualified path of the package}
}
\value{
string Relative file path where packages of
this type should be copied to.
}
\description{
This function returns the directory path (relative to
the repo root) where the package needs to be copied to.
}
\author{
Jan Schulz, Dirk Eddelbuettel and Matthew Jones
}

