% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insertPackage.R
\name{getPathForPackageType}
\alias{getPathForPackageType}
\title{Get relative path for package type}
\usage{
getPathForPackageType(pkgtype, rversion)
}
\arguments{
\item{pkgtype}{The package type as a string.}

\item{rversion}{String which identifies the major.minor
to the version of the current interpreter.}
}
\value{
string Relative file path where packages of
this type should be copied to.
}
\description{
This function returns the directory path (relative to
the repo root) where the package needs to be copied to.
}
\author{
Jan Schulz and Dirk Eddelbuettel
}

