% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertPackage.R
\name{identifyPackageType}
\alias{identifyPackageType}
\title{Identifies the package type from a filename}
\usage{
identifyPackageType(file, pkginfo = getPackageInfo(file))
}
\arguments{
\item{file}{An R package in source or binary format,}

\item{pkginfo}{information on the R package referenced by \code{file}}
}
\value{
string Type of the supplied package.
}
\description{
This function identifies the package type from a filename.
}
\details{
The returned string is suitable for \code{write_PACKAGES()}.
}
\author{
Jan Schulz and Dirk Eddelbuettel
}
