% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initRepo.R
\name{initRepo}
\alias{initRepo}
\title{Intialize a git repo for drat}
\usage{
initRepo(name = "drat", basepath = "~/git", remote_type = "SSH",
  push = FALSE)
}
\arguments{
\item{name}{A character variable with the name the new repository, the
default is \dQuote{drat}. If name has two parts in format of
\dQuote{account_name/repo_name}, the cooresponding github repo is set as
repo git remote. The initialized repo is pushed to remote in the end.}

\item{basepath}{A character variable with path to the directory in which the
new repository is to be created. The default value is \dQuote{~/git}.}

\item{remote_type}{Add github remote with \dQuote{SSH} or \dQuote{HTTPS}
format.}

\item{push}{Push changes to github remote in the end if available. This may
only works for SSH and when ssh-agent is available.}
}
\value{
The function is invoked for its side-effects and only returns
  \code{NULL} invisibly.
}
\description{
This helper function create a new repository, creates and checks out
\sQuote{gh-pages} branch and fills it with the required path for both source
and binary repositories. The initialized repo is pushed to github remote in
the end if \sQuote{name} includes both account name and repo name.
}
\details{
This function is still undergoing development and polish and may change in
subsequent versions.
}
\author{
Dirk Eddelbuettel
}
