% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Insert a source archive into a drat repository}
\usage{
insert(file, repodir = getOption("dratRepo", "~/git/drat"), commit = FALSE)
}
\arguments{
\item{file}{A source archive in \code{tar.gz} form.}

\item{repodir}{A local directory corresponding to the repository
top-leve directory.}

\item{commit}{A boolean toggle to select automatic git operations
\sQuote{add}, \sQuote{commit}, and \sQuote{push}.}
}
\value{
NULL is returned.
}
\description{
R can use multiple archives: CRAN, BioConductor and Omegahat have
been supported for years.  It is equally easy to add local
archives from the same machine, or local network, or university /
company network as well as other publically available
repositories.  This function aids in the process, and defaults to
inserting a given source archive into a given repository.
}
\details{
This function inserts the given file into the given (local)
repository and updates the index. One can then push to remote
such as GitHub.
}
\author{
Dirk Eddelbuettel
}

