% Generated by roxygen2: do not edit by hand
\name{insertArtifacts}
\alias{insertArtifacts}
\title{Generate R package artifacts after Build and Check}
\usage{
insertArtifacts(file = list.files(pattern = "*\\\\.tar\\\\.gz$"),
  repodir = "public", repo.url = character(), repo.cran = FALSE,
  log.files = character())
}
\arguments{
\item{file}{An R package in source or binary format.}

\item{repodir}{A local directory corresponding to the repository top-level directory.}

\item{repo.url}{character, url to be used in \emph{repos} argument in generated \code{install.packages} call for html website.}

\item{repo.cran}{logical or character, default FALSE, if TRUE it will append repository list with CRAN repo in generated \code{install.packages} call.}

\item{log.files}{character vector of files to copy from \emph{*.Rcheck} dir, for drat that could be \code{tests/simpleTests.Rout}.}
}
\value{
TRUE invisibly if \code{index.html} generated successfully.
}
\description{
Generate basic html website with basic package info, includes html manuals, package installation string, links to check logs. Additionally \emph{CI} environment variables for git branch and git commit will be included if available.
}
\details{
Function assumes the working directory contains package source tar.gz file made by \code{R CMD build} and package \emph{.Rcheck} directory made by \code{R CMD check}. To produce html manual it will install the package to temporary library.
}
\examples{
\dontrun{
  insertArtifacts(file = "drat_0.1.0.1.tar.gz",
                  repodir = "drat",
                  repo.url = "https://eddelbuettel.github.io/drat")
}
}
\author{
Jan Gorecki
}

