% Generated by roxygen2: do not edit by hand
\name{packageHtml}
\alias{packageHtml}
\title{Generate basic html website based on package metadata}
\usage{
packageHtml(file = list.files(pattern = "*\\\\.tar\\\\.gz$"), repodir = ".",
  repo.url = character(), repo.cran = FALSE, manual = FALSE)
}
\arguments{
\item{file}{An R package in source or binary format.}

\item{repodir}{A local directory corresponding to the repository top-level directory.}

\item{repo.url}{character, url to be used in \emph{repos} argument in generated \code{install.packages} call for html website.}

\item{repo.cran}{logical or character, default FALSE, if TRUE it will append repository list with CRAN repo in generated \code{install.packages} call.}

\item{manual}{logical, default FALSE, when TRUE it will copy package \emph{html} subdirectory to \emph{repodir} and add a link to \code{html/00Index.html}. Additionally \emph{DESCRIPTION} file will be copied so can be accessed from manual.}
}
\value{
TRUE is returned.
}
\description{
Generate basic html website with basic package info, optionally includes html manual. Package installation string can be included as well. \emph{CI} environment variables for git branch and git commit will be included if available.
}
\examples{
\dontrun{
  packageHtml(file = "drat_0.1.0.tar.gz",
              repodir = "drat",
              repo.url = "https://eddelbuettel.github.io/drat",
              manual = TRUE)
}
}
\author{
Jan Gorecki
}

