% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/archivePackages.R
\name{archivePackages}
\alias{archivePackages}
\title{Move older copies of packages to an archive}
\usage{
archivePackages(repopath = getOption("dratRepo", "~/git/drat"),
  type = "source", pkg)
}
\arguments{
\item{repopath}{Character variable with the path to the repo;
defaults to the value of the \dQuote{dratRepo} option with
\dQuote{"~/git/drat"} as fallback}

\item{type}{Character variable for the type of repository, so far \dQuote{source}}

\item{pkg}{Optional character variable specifying a package name(s), whose
older versions should be archived. If missing (the default), archiving is
performed on all packages.}
}
\description{
The function moves older versions of packages into a CRAN-style
archive folder.
}
\details{
This function is still undergoing development and polish and may
change in subsequent versions.
}
\examples{
\dontrun{
  archivePackages()   # archive all older package versions
  archivePackages(pkg = "drat")  # archive older copies of just one package
}
}
\author{
Thomas J. Leeper
}

