% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertPackage.R
\name{getPackageInfo}
\alias{getPackageInfo}
\title{Get information from a binary package}
\usage{
getPackageInfo(file)
}
\arguments{
\item{file}{the fully qualified path of the package}
}
\value{
A named vector with several components
}
\description{
This function returns the compile-time information added
to the \code{DESCRIPTION} file in the package.
}
\section{Note}{

This is an internal function, use \code{:::} to access it from outside
the internal package code.
}

\author{
Dirk Eddelbuettel
}
