% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insertPackage.R
\name{identifyPackageType}
\alias{identifyPackageType}
\title{Identifies the package type from a filename}
\usage{
identifyPackageType(file)
}
\arguments{
\item{file}{An R package in source or binary format,}
}
\value{
string Type of the supplied package.
}
\description{
This function identifies the package type from a filename.
}
\details{
The returned string is suitable for \code{write_PACKAGES()}.
}
\author{
Jan Schulz and Dirk Eddelbuettel
}

