% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initRepo.R
\name{initRepo}
\alias{initRepo}
\title{Intialize a git repo for drat}
\usage{
initRepo(name = "drat", basepath = "~/git")
}
\arguments{
\item{name}{A character variable with the name the new repository,
the default is \dQuote{drat}.}

\item{basepath}{A character variable with path to the directory in
which the new repository is to be created. The default value is
\dQuote{"~/git"}.}
}
\value{
The function is invoked for its side-effects and only
returns \code{NULL} invisibly.
}
\description{
This helper function create a new repository, creates and checks
out \sQuote{gh-pages} branch and fills it with the required path.
}
\details{
Currently only \sQuote{src/contrib} for source repositories is
supported.

This function is still undergoing development and polish and may
change in subsequent versions.
}
\author{
Dirk Eddelbuettel
}

